requirements.txt
Flask==3.1.0
Flask-SocketIO==5.3.6
Flask-CORS==5.0.1
python-chess
eventlet>=0.36.1
gunicorn==21.2.0 